MNT: Use float instead of np.float, int instead of np.int

+++ ginga/rv/plugins/RC.py
@@ -186,8 +186,6 @@ import sys
 import bz2
 from io import BytesIO
 
-import numpy as np
-
 from ginga import GingaPlugin
 from ginga import AstroImage
 from ginga.gw import Widgets
@@ -381,7 +379,7 @@ class GingaWrapper(object):
 
             # dtype string works for most instances
             if dtype == '':
-                dtype = np.float
+                dtype = float
 
             byteswap = metadata.get('byteswap', False)
